refactor:move private functions to utils.go of export package
